requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-chroma
chromadb
pypdf
python-dotenv
langchain-openai
langchain-anthropic
langchain-ollama
crawl4ai
beautifulsoup4
torch
langchain-huggingface
sentence-transformers
